Fix a subtle problem in Appender processing in the appender.getTrafficClusters() (#7689)
Fix a subtle problem in Appender processing in the appender.getTrafficClusters() 
utility function.  This function depended on the "isInaccessible" node
metadata being set. But "isInaccessible" is set in the "Outsider" appender,
a Finalizer which is run late in the appender processing, so appenders using
this utility prior to the Outsider Finalizer being run, were at risk.

Update appender.getTrafficClusters() to not depend on that node metadata
being set.