[Ambient] Add bi-directional edge support for waypoint traffic (#7719)
This PR collapses two edges into one when there is bi-directional traffic to-from a node and a waypoint.  This simplifies the graph significantly.  This PR provides initial side-panel support for the bidirectional edges, although UX is ongoing and there may still be revisions.  In short, the approach is to "hide" the "from" edge as metadata on the "to" edge.