When control plane information was shown in a namespace card on the Overview page, it was included as part of namespace metrics. But now that we have more fine-grained control plane nodes in the Mesh page, and there may be multiple control planes, we need to move the logic from namespace metrics and ensure that the retrieved information is specific to the specific control place workload.
The new logic creates dedicated control plane metric support. It determines
the pods for the specific control plane deployment, and uses the pod names
to narrow metrics to just the relevant time series.

Also, I think there was a flaw in the prior logic for determining the proxy
sync time. Those metrics are unusual as they are infrequent and reported
more like a step function. You can't really capture it well as a
"rate" because it reports like a step function, all at once when there is
a config push, and then goes flat.  Rates need a series of data points. So,
this applies some non-standard logic to capture any sync that takes place
during the requested time period, by capturing "deltas".  This turned out to
be a bit painful, as we didn't really have the underlying support for the
type of prom queries needed. So, this adds some support for that.