Protect against minigraph problems when the parent detail page component is changing its target element.
This is a semi-hack to prevent a major refactoring of how we work with
the GraphDatasource object in the detail pages. Because we re-use the
GraphDatasource, and it's an asynch fetch, it's possible that if the
component gets re-used for a new workload/service/app, that things get
out of sync.

In short, detect a mismatch between the target component and the datasource,
and avoid updating the graph until things get back in sync.

Also:
- remove an unwanted dependency on "graphData" in effect that responds to
  a change in selectedIds
- slightly shorten the focus node flashing.
- suppress a bunch of console warnings.

https://github.com/kiali/kiali/issues/7839