PFT Graph and Mesh hide had needed to call e.remove() to detach hidden (#7885)
edges from their parent group. I had inline comments stating that it
really shouldn't be necessary. Now, either do to a PFT fix, and/or
our recent change to wrap element updates in mobx actions, it seems
unnecessary, and seems to be causing this orphan edge issue. Which
makes sense because we're detaching it from its parent.

https://github.com/kiali/kiali/issues/7878