For completeness, we should also allow users the ability to set the (#7923)
trace fetch limit in the graph's node-side-panel Traces tab.

As part of this, remove the "use graph refresh" option and just have
users manually refresh the traces list, as needed. This simplifies
the tab header, given the new dropdown, and I doubt anyone will miss
this option. If they do, we can always bring it back.

part-of kiali7769