Support "canonical" handling of app and version labels (#8122)
Historically, Kiali supports a single set of app/version label names and the Kiali defaults to the original Istio label names:
  app, version.

Over time the labeling schemes have evolved, and many users prefer the k8s label names of:
  app.kubernetes.io/name, app.kubernetes.io/version

And to handle this mixed approach in telemetry, Istio introduced the idea of a canonical app and version, and with it introduced the label names:
  service.istio.io/canonical-name, service.istio.io/canonical-revision

When setting the canonical value, the order of preference is:
  1. service.istio.io/canonical-name, service.istio.io/canonical-revision
  2. app.kubernetes.io/name, app.kubernetes.io/version
  3. app, version

What has resulted for many users is a mixed labeling scheme. And as such, Kiali's single scheme, as defined in the Kiali CR, is limiting to those users, where certain objects may be missed.

This relieves that limitation by allowing for a mixed scheme by default. The Kiali CR now defaults to unset values for istio_labels.app_label_name and istio_labels.version_label_name, and will accept any of the three schemes above. For users with a consistent scheme, they can still set the CR to their scheme (one of the above, or something custom).

Kiali does not support a mix of app and version label names. For example, if labeling a deployment with app, it is assumed that the version label name would be version. Although, it is valid to label the same object with multiple schemes. The same service could be labeled with both app=foo and app.kubernetes.io/name=foo. Note that when determining the app label value used on the service, the canonical order of preference is applied. So, in the [unrecommended] situation where the same service is labeled with both app=foo and app.kubernetes.io/name=bar, bar would be the preferred value.