[perf] validation perf work (#8194)
This PR does not fundamentally change the way we do validations. The general approach to the refactor was to "pull up" anything that could be calculated or fetched one time, and then re-used. The existing code buried a fair amount of calculation in functions that were called many times in one reconciliation, as there is a lot of looping going on. Despite caching, the repeated lookup and calculation can add up in larger deployments.

This introduces an validationInfo object, vInfo, that accumulates information used for the validation pass. It gets allocated up front via a NewValidationInfo call, and has three parts: base, cluster, namespace. Base information can be calculated one time is stored in the initial New. Cluster information must be reset for each cluster being processed, and Namespace info for each Namespace. The vInfo object is then passed around throughout the process, and used to define the checkers. Note, to minimize changes I stopped short of passing it into the checkers themselves.