Reduce calls to config.Get() (#8227)
This large refactor doesn't fundamentally change the use of the Kiali config; the goal is to reduce pressure on the config's read lock and eliminate a large amount of copy/garbage-collection of the config data.

This is motivated by the work to reduce the footprint for running validations, where pprof results indicated high amounts of GC, but the PR extends across the code-base. In short, calls to config.Get() have been pulled up in the code and the config is passed down to the lower-level code that needs it. For example, because the business services embed the config, you will no longer find calls to config.Get() in any of the business service API impls.

Other:
- use "conf" variable name consistently, to avoid overlap with the config package name and for clarity.
- almost always pass conf by reference
- ensure conf is available in every business service