Cytoscape removal (part 1) (#8252)
This PR is part 1 of ending the deprecation of the Cytoscape graph impl. It removes all cytoscape and moves files around.  Part 2 will do some file renaming and some other cleanup.  It is done i two parts to better show the file change history.

On the UI side, it:
- removes the cytoscape graph impl
- removes the cytoscape package dependencies
- removes the related unit and cypress tests
- converts the graph find/hide unit tests from Cy selectors to our PFT elem "selector" utils
- simplifies the layout handling by
- removing layout classes
- making trafficGraph and Mesh layout definitions distinct
- fixes some URL param handling by ensuring a valid layout value
- consolidates all [pft] traffic graph impl under src/pages/graph
- remove PFColorValues, which were only needed for cytoscape
also:
I ran into an odd issue with an enum export not available when running yarn test. I had to relocate the enum, I think to avoid some sort of circular dependency.

On the server side, it:
- removes the related configuration (also see operator PR)
- converts graph/config/cytoscape package to graph/config/common
  - this reflects that we only have a single, "common" graph config format provided to clients. Although, the arch still 
    supports the ability to provide different config vendors.

kiali8053