update PFT cypress tests with some logic in the cytsoscape cypress  tests (#8270)
- In a PR for issue 7256, and the subsequent PR https://github.com/kiali/kiali/pull/7442,
we updated cytoscape cypress tests to more reliably wait for the graph
to be populated before testing for edges and nodes. Probably due to timing,
those changes were not applied to the analogous PFT cypress tests.

This PR applies those changes, ensuring the data "isLoading=false", and enhances them
to also ensure the graph "isReady=true".
- Add "isLoading" state to MiniGraphCard so that we can check both states. For the minigraph these
  two things are handled slightly differently than in the main graph.