remove GraphSelectorBuilder (#8334)
* This removes GraphSelectorBuilder, which was implemented based on Cytoscape
Selector syntax. It allowed non-graph code, like Trace Detail to navigate
to the graph and attempt to focus on the related node (perform the
node "flashing"), to help orient the user the relevant node on the graph.

In the PFT graph we only offer node focus based on NodeId. We could have
implemented something similar for PFT, using our "home grown selectors",
but it was decided (well, mainly by me) that the slight regression
didn't merit the effort. I don't think users will notice, and they always
have the option to use Graph Find to locate the node, if desired.

- also, replaced some 'focusSelector' literals with an existing constant.

kiali8258