add structured logging with context (#8363)
* add structured logging with context

* here's a sample of adding a "group" name (for lack of a better term)
to the logger so when we log a message there will be context that
includes the group name (in this case "kialiCache")

* adding http request logging infrastructure
adding Config handler log message as an initial test to see what it looks like

* abbreviate the names of message and timestamp keys to save chars in the json output. this also follows the slog convention, too.

* upgrade zerolog to latest version

* add all context for the request (for now).
This logs everything that hlog will know about the request.
I put them all in here because it will be easier to remove
what we don't want later. This at least will show us everything
in the logs and we can determine what we like or don't like here.

* add a debug message when apps in a cluster are requested.
This is just to see what the structured log looks like for
a request that has params in the URL

* remove all but request id
rename RequestID

* resolving conflicts and updating new log code that was added to use the new group logger