more logging infrastructure (#8391)
* more group logging - found way to retain "log." so some code can remain as-is

* this is really annoying - not just passing around a logger, but passing around *zerolog.Logger
when in other places we are passing around our own log.ContextLogger.

Starting to think we need to just bit the bullet and move all logging API to all zerolog so we aren't using all different logger types all over.

* move to using zerolog API directly

* now shows request ID in the logs

* guardrail against nil ptr - no context just means start with an empty one

* zerolog should be a direct dep

* pass context around to get the logger

* log panic stack trace using context logger

* remove commented out method - we don't need it

* fix name of constant

* revert controller.go - the Start method is going away and nick will make that change. I'll leave it alone for this PR.

* revert changes to validator - not sure best way to do this so saving that for later

* pass context instead, since that is what we really need

* 1. get the logger correctly configured in the graph generator calls right at the start
2. remove logger.go files and put all log name constants in a common file in log package

* add convienence function to allow our handler code to set a new group in a new logger in a request context.
The call chain to do this is long and complicated and error prone, so this function is nice to make things easier.

* store logger directly
in the future, we might want to get rid of this, too, and just pass the logger inside a context to all function calls