[CI] skip molecule tests with OLM-installed operator in release workflow
For release workflow, we do not want to run molecule tests with
the operator installed via OLM because OperatorHub.io will not have
the new operator being released. OperatorHub.io will only have the
last release and it might not pass the tests meant for the new
version of the operator that is about to be released.