Forward x-request-id header to external calls to prometheus (#8767)
* Forward x-request-id header to external calls to prometheus

fixes: https://github.com/kiali/kiali/issues/8468

* this commit passes context to the prometheus client.
this will enable http tracing when the prom client is called from within
the context of an HTTP request. If the prom client is not called from
within an HTTP request, then it will fallback to using the context
that was given to the client at instantiation time. This is usedful when
non-HTTP request jobs use the prom client but we still want some good logging for those
calls. But what this code now does is provide prometheus client tracing
within HTTP requests - so, for example, we can capture this data in
Tempo and see Kiali trace logs down to the prom client calls.

* i don't know why I need the port for kind now, before I did not