Add the ability for kiosk parents to supply parent-specific kiosk data (#8819)
Add the ability for parents to supply parent-specific kiosk data, to support
conditional behavior in the core code. This is a little messy, as we don't want
extensive conditional logic in the core code. But, we already have messy handling
that this may help start a clean up, and there are times when we may need to know
more about what the parent can support than a simple string can easily allow for.