[PF6] fix graph label issues (#8973)
* To deal with the label icon coloring, I originally thought that we needed to
explicitly react to a 'theme' change. But in the end, with PF6 we can now
defer to PFTopo to handle this for us. So, just remove the explicit coloring
for the icon.

* The box label width issue had to do with our style override for hiding the kebab. This
has been fixed, so the issue goes away when using the fixed mechanism.

Also, remove the the group label font size override. It wasn't working and actually,
when I applied it explicitly (using labelClassName in StyleGroup), it didn't look good.
Better to have consistent label sizes.

I did keep the increased MeshNode font, and added the equivalent change to MeshGroup.
This looks better for the less dense mesh graphs.

* an unrelated change, but something lingering from the PFT migration, remove the
extraneous `()` for namespace and/or cluster,  in the secondary labels.

* tweak shortcuts

* Improve the badge readability. I don't think we are playing with the group label
background color, it seems PFT is now using a "reverse" color approach given the
theme. Using the default badge colors doesn't look very clear, so I chose colors that
seem to work on each theme.

* Adjust group padding and label badge styles

* Revert change in group padding, I think the symmetrical padding is better.

---------

Co-authored-by: Fernando Hoyos <fhoyosle@redhat.com>